Adjust code to quieten warnings

+++ tests/test_assert.c
@@ -76,14 +76,14 @@ int test_message(void* p) {
 }
 /* test message output : %% */
 int test_message_percent(void* p) {
-  int bytes = reedkiln_log_printf("100%%");
+  size_t bytes = reedkiln_log_printf("100%%");
   reedkiln_assert(bytes == 5);
   return Reedkiln_OK;
 }
 /* test message output : %s */
 int test_message_s(void* p) {
   char const str[] = "Hello, world!";
-  int bytes = reedkiln_log_printf("Quotation: \"%s\"", str);
+  size_t bytes = reedkiln_log_printf("Quotation: \"%s\"", str);
   reedkiln_assert(bytes == 13+14);
   return Reedkiln_OK;
 }
@@ -91,7 +91,7 @@ int test_message_s(void* p) {
 int test_message_s_precision(void* p) {
   char const str[] = "abcdefghijklm";
   unsigned int const num = reedkiln_rand()%12;
-  int bytes = reedkiln_log_printf("Quotation: \"%.*s\" %.5s",
+  size_t bytes = reedkiln_log_printf("Quotation: \"%.*s\" %.5s",
     num, str, str);
   reedkiln_assert(bytes == 15+num+5);
   return Reedkiln_OK;
@@ -100,7 +100,7 @@ int test_message_s_precision(void* p) {
 int test_message_ls(void* p) {
   wchar_t const str[] = L"Hello, world!";
   wchar_t const str2[] = { 0x101, 0x250C, 0 };
-  int bytes = reedkiln_log_printf("Wide %ls: ", str);
+  size_t bytes = reedkiln_log_printf("Wide %ls: ", str);
   reedkiln_log_printf("%ls", str2);
   reedkiln_assert(bytes == 8+13);
   return Reedkiln_OK;
@@ -109,15 +109,15 @@ int test_message_ls(void* p) {
 int test_message_ls_precision(void* p) {
   wchar_t const str[] = L"abcdefghijklm";
   unsigned int const num = reedkiln_rand()%12;
-  int bytes = reedkiln_log_printf("Wide %.*ls", num, str);
+  size_t bytes = reedkiln_log_printf("Wide %.*ls", num, str);
   reedkiln_assert(bytes == 6+num);
   return Reedkiln_OK;
 }
 /* test message output : %x */
 int test_message_x(void* p) {
   unsigned int num = reedkiln_rand();
-  int bytes = reedkiln_log_printf("Number %x.", num);
-  int expected = 0;
+  size_t bytes = reedkiln_log_printf("Number %x.", num);
+  size_t expected = 0;
   if (num == 0)
     expected = 1;
   else while (num > 0) {
@@ -129,20 +129,20 @@ int test_message_x(void* p) {
 }
 /* test message output : %p */
 int test_message_p(void* p) {
-  int bytes = reedkiln_log_printf("Pointer %p != %p.", &p, NULL);
+  size_t bytes = reedkiln_log_printf("Pointer %p != %p.", &p, NULL);
   reedkiln_assert(bytes > 10);
   return Reedkiln_OK;
 }
 /* test message output : %a */
 int test_message_a(void* p) {
-  int bytes = reedkiln_log_printf("Float %a %a %a %a.",
+  size_t bytes = reedkiln_log_printf("Float %a %a %a %a.",
     1.0, 0.0, -1.0, HUGE_VAL);
   reedkiln_assert(bytes >= 10+7);
   return Reedkiln_OK;
 }
 /* test message output : %a at the limits */
 int test_message_a_limits(void* p) {
-  int bytes = reedkiln_log_printf("Float %a %a %a %a.",
+  size_t bytes = reedkiln_log_printf("Float %a %a %a %a.",
     DBL_MAX, DBL_MIN, DBL_EPSILON, FLT_EPSILON);
   reedkiln_assert(bytes >= 10+7);
   return Reedkiln_OK;
@@ -152,7 +152,7 @@ int test_message_a_rand(void* p) {
   unsigned int num = reedkiln_rand();
   double v = ((num&1) ? -1.0 : 1.0) * (num * 256.0 / UINT_MAX);
   double w = ((num&1) ? -1.0 : 1.0) * (num / 256.0 / UINT_MAX);
-  int bytes = reedkiln_log_printf("%a * %a = %a;", v, w, v*w);
+  size_t bytes = reedkiln_log_printf("%a * %a = %a;", v, w, v*w);
   reedkiln_assert(bytes > 10);
   return Reedkiln_OK;
 }
@@ -161,7 +161,7 @@ int test_message_a_rand(void* p) {
 int test_message_i(void* p) {
   unsigned int num = reedkiln_rand()%INT_MAX;
   unsigned int const old_num = num;
-  int bytes = reedkiln_log_printf("Number %i.", old_num);
+  size_t bytes = reedkiln_log_printf("Number %i.", old_num);
   int expected = 0;
   if (num == 0)
     expected = 1;
@@ -177,7 +177,7 @@ int test_message_x_precision(void* p) {
   size_t const positive = ((~(size_t)0)>>1);
   size_t num = (reedkiln_rand()*(size_t)12345) & positive;
   size_t const old_num = num;
-  int bytes = reedkiln_log_printf("Number %0zX.", old_num);
+  size_t bytes = reedkiln_log_printf("Number %0zX.", old_num);
   int expected = 0;
   if (num == 0)
     expected = 1;
@@ -193,8 +193,8 @@ int test_message_i_precision(void* p) {
   size_t const positive = ((~(size_t)0)>>1);
   size_t num = (reedkiln_rand()*(size_t)12345) & positive;
   size_t const old_num = num;
-  int bytes = reedkiln_log_printf("Number %*.*zi.", 1, 1, old_num);
-  int expected = 0;
+  size_t bytes = reedkiln_log_printf("Number %*.*zi.", 1, 1, old_num);
+  size_t expected = 0;
   if (num == 0)
     expected = 1;
   else while (num > 0) {
@@ -207,8 +207,8 @@ int test_message_i_precision(void* p) {
 /* test message output : %i negative */
 int test_message_i_negative(void* p) {
   int const old_num = -(int)(reedkiln_rand()%INT_MAX);
-  int bytes = reedkiln_log_printf("Number %i.", old_num);
-  int expected = 0;
+  size_t bytes = reedkiln_log_printf("Number %i.", old_num);
+  size_t expected = 0;
   int num = -old_num;
   if (num == 0)
     expected = 0;
